requirements.txt
python-binance
ta-lib
pandas
numpy
websocket-client
pprintpp
mplfinance
datetime
